refactor(utils): split module into submodules
