+++ src/settings/knn_classifier_parameters.rs
@@ -1,6 +1,6 @@
 //! KNN classifier parameters
 
-use crate::utils::Distance;
+use crate::utils::distance::Distance;
 pub use smartcore::{algorithm::neighbour::KNNAlgorithmName, neighbors::KNNWeightFunction};
 
 /// Parameters for k-nearest neighbors (KNN) classification
